move view override to trait
